feat: move TextContent
