+++ packages/vrender-components/src/tooltip/type.ts
@@ -6,7 +6,6 @@ import type {
   IRichTextCharacter
 } from '@visactor/vrender-core';
 import type { IBounds, IPadding } from '@visactor/vutils';
-import type { TextContent } from '../tag';
 
 export interface IContainerSize {
   width?: number;
